feat(projectsDetail): add DocumentsCard component for file display and management
