+++ apps/public/app/[lang]/(landing)/projects/[id]/client.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import {Avatar, AvatarFallback} from "@/components/ui/avatar";
-import {Button} from "@/components/ui/button";
 import {Card, CardContent, CardHeader, CardTitle} from "@/components/ui/card";
 import {toast} from "@/components/ui/sonner";
 import type {
@@ -16,6 +15,7 @@ import type {JSONContent} from "@repo/ayasofyazilim-ui/organisms/tiptap";
 import TipTapEditor from "@repo/ayasofyazilim-ui/organisms/tiptap";
 import {ChevronRight} from "lucide-react";
 import Link from "next/link";
+import Image from "next/image";
 import {useParams} from "next/navigation";
 import React, {useCallback, useState} from "react";
 import {useMember} from "@/app/providers/member";
@@ -24,38 +24,7 @@ import FundingTable from "../_components/funding-card";
 import MobileSupportDrawer from "../_components/mobile-support-card";
 import ProjectSummary from "../_components/project-summary";
 import ProjectActions from "./_components/project-actions";
-
-// Add this mock data at the top of the file, after the imports
-const mockDocuments = [
-  {
-    id: 1,
-    title: "Proje Başvuru Belgesi",
-    description: "Projenin resmi başvuru ve onay belgesi",
-    date: "2023-12-15",
-    type: "pdf",
-  },
-  {
-    id: 2,
-    title: "Patent Sertifikası",
-    description: "Teknoloji için alınan uluslararası patent belgesi",
-    date: "2024-01-20",
-    type: "pdf",
-  },
-  {
-    id: 3,
-    title: "Ödül Sertifikası - Medikal İnovasyon 2023",
-    description: "Uluslararası Medikal İnovasyon Yarışması Birincilik Ödülü",
-    date: "2023-11-05",
-    type: "image",
-  },
-  {
-    id: 4,
-    title: "Hukuki Durum Raporu",
-    description: "Projenin güncel yasal durumu ve fikri mülkiyet hakları raporu",
-    date: "2024-02-10",
-    type: "doc",
-  },
-];
+import DocumentsCard from "./_components/documents-card";
 
 export default function ProjectDetails({
   data,
@@ -108,66 +77,6 @@ export default function ProjectDetails({
       toast.error("An error occurred while processing your payment");
     } finally {
       setIsLoading(false);
-    }
-  };
-
-  const getFileIcon = (type: string) => {
-    switch (type) {
-      case "pdf":
-        return (
-          <svg
-            className="lucide lucide-file-text"
-            fill="none"
-            height="20"
-            stroke="currentColor"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth="2"
-            viewBox="0 0 24 24"
-            width="20"
-            xmlns="http://www.w3.org/2000/svg">
-            <path d="M14.5 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V7.5L14.5 2z" />
-            <polyline points="14 2 14 8 20 8" />
-            <line x1="16" x2="8" y1="13" y2="13" />
-            <line x1="16" x2="8" y1="17" y2="17" />
-            <line x1="10" x2="8" y1="9" y2="9" />
-          </svg>
-        );
-      case "image":
-        return (
-          <svg
-            className="lucide lucide-image"
-            fill="none"
-            height="20"
-            stroke="currentColor"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth="2"
-            viewBox="0 0 24 24"
-            width="20"
-            xmlns="http://www.w3.org/2000/svg">
-            <rect height="18" rx="2" ry="2" width="18" x="3" y="3" />
-            <circle cx="9" cy="9" r="2" />
-            <path d="m21 15-3.086-3.086a2 2 0 0 0-2.828 0L6 21" />
-          </svg>
-        );
-      default:
-        return (
-          <svg
-            className="lucide lucide-file"
-            fill="none"
-            height="20"
-            stroke="currentColor"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth="2"
-            viewBox="0 0 24 24"
-            width="20"
-            xmlns="http://www.w3.org/2000/svg">
-            <path d="M14.5 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V7.5L14.5 2z" />
-            <polyline points="14 2 14 8 20 8" />
-          </svg>
-        );
     }
   };
 
@@ -309,28 +218,28 @@ export default function ProjectDetails({
                   {(paymentResponse.totalCount || 0) > 3 && (
                     <div className="bg-background flex items-center rounded-full border p-1 shadow-sm">
                       <div className="flex -space-x-1.5">
-                        <img
+                        <Image
                           alt="Avatar 01"
                           className="ring-background rounded-full ring-1"
                           height={20}
                           src="https://originui.com/avatar-80-03.jpg"
                           width={20}
                         />
-                        <img
+                        <Image
                           alt="Avatar 02"
                           className="ring-background rounded-full ring-1"
                           height={20}
                           src="https://originui.com/avatar-80-04.jpg"
                           width={20}
                         />
-                        <img
+                        <Image
                           alt="Avatar 03"
                           className="ring-background rounded-full ring-1"
                           height={20}
                           src="https://originui.com/avatar-80-05.jpg"
                           width={20}
                         />
-                        <img
+                        <Image
                           alt="Avatar 04"
                           className="ring-background rounded-full ring-1"
                           height={20}
@@ -351,56 +260,7 @@ export default function ProjectDetails({
               </CardContent>
             </Card>
           )}
-          <Card className="mt-6">
-            <CardHeader>
-              <CardTitle className="text-xl font-bold md:text-2xl">Belge, Ödül, Hukuki Durum</CardTitle>
-            </CardHeader>
-            <CardContent>
-              <div className="space-y-4">
-                {mockDocuments.map((doc) => (
-                  <div
-                    className="hover:bg-muted/50 group flex items-center gap-3 rounded-lg border p-3 transition-all"
-                    key={doc.id}>
-                    <div className="bg-primary/10 text-primary flex h-10 w-10 shrink-0 items-center justify-center rounded-md">
-                      {getFileIcon(doc.type)}
-                    </div>
-                    <div className="flex-1">
-                      <div className="flex items-center justify-between">
-                        <h4 className="text-sm font-medium">{doc.title}</h4>
-                        <Button
-                          className="h-8 w-8 opacity-0 transition-opacity group-hover:opacity-100"
-                          size="icon"
-                          variant="ghost">
-                          <svg
-                            className="lucide lucide-download"
-                            fill="none"
-                            height="16"
-                            stroke="currentColor"
-                            strokeLinecap="round"
-                            strokeLinejoin="round"
-                            strokeWidth="2"
-                            viewBox="0 0 24 24"
-                            width="16"
-                            xmlns="http://www.w3.org/2000/svg">
-                            <path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4" />
-                            <polyline points="7 10 12 15 17 10" />
-                            <line x1="12" x2="12" y1="15" y2="3" />
-                          </svg>
-                          <span className="sr-only">İndir</span>
-                        </Button>
-                      </div>
-                      <p className="text-muted-foreground line-clamp-2 text-sm">{doc.description}</p>
-                      <div className="mt-1 flex items-center gap-2">
-                        <span className="text-muted-foreground text-xs">{doc.date}</span>
-                        <span className="bg-muted-foreground inline-flex h-1.5 w-1.5 rounded-full" />
-                        <span className="text-muted-foreground text-xs uppercase">{doc.type}</span>
-                      </div>
-                    </div>
-                  </div>
-                ))}
-              </div>
-            </CardContent>
-          </Card>
+          <DocumentsCard fileResponse={fileResponse} />
         </div>
       </div>
     </main>
