+++ apps/public/next.config.mjs
@@ -12,6 +12,10 @@ const nextConfig = {
         protocol: "https",
         hostname: "s3.us-west-2.amazonaws.com",
       },
+      {
+        protocol: "https",
+        hostname: "originui.com"
+      }
     ],
   },
 };
